RegTAP: now requiring security_method_id to be lowercased.
This is a bit of a conundrum, because it will kill retrievability of such
URLs if they contain uppercase characters in the path part of the URL and
the scheme doesn't that's case-insensitive.

On the other hand, most of these will be ivoids, and these must be matched 
case-insensitively.  Ah well.

Also, several editorial changes in table descriptions up to the interface
table.